B: split long property values into multiple chunks
concourse/concourse#7284

Our containerd runtime started erroring when resource versions would
emit metadata strings that were too long. This is because we store the
version + metadata (encoded as JSON) on the container as a
garden.Property after the `in`/`out` script succeeds (so that if the ATC
re-attaches to the build, we can check if we've already run the script,
and if so, retrieve the result).

We previously stored garden properties as containerd labels in a 1:1
mapping. However, containerd v1.5.3 started rejecting labels that were
"too long" (i.e. where the length of the key + value exceeded 4096
bytes).

As a workaround, we now split property values up into multiple chunks
that get stored as sequential labels as needed. For instance, suppose we
wanted to save the following property:

foo: aaaa....(5000 times)

This will be stored as the following labels in containerd:

foo.0: aaaa....(4091 times, so that the len(key) + len(value) == 4096)
foo.1: aaaa....(909 times)

Note that if a key + value does not exceed the limit, we still use the
same `key.0` pattern. e.g.

foo: aaa

is stored as

foo.0: aaa

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>
Co-authored-by: Esteban Foronda <eforonda@vmware.com>