testflight: remove check in resource type defaults test
The test for checking if the defaults are propagated to resource sources
for check step was flaking because now that we no longer error when the
parent resource type does not have a version, the initial interval check
that happens once we unpause the pipeline goes through and updates the
last check start time. This becomes a problem when that time that it
updates the last check start time is after the time that we also manuall
trigger a check resource for the test. If the last check start time is
after the time that we manually check the resource, the manual check
will not run (https://github.com/concourse/concourse/pull/7239). This is
as intended so I think we can just remove checking for the defaults for
a check step as it is tested in unit tests and also we will be checking
for the defaults in the get and put steps.

[#6274]

Signed-off-by: Clara Fu <fclara@vmware.com>