speed up TSA tests
The HTTP client used by the TSA uses http.DefaultTransport, which has an
IdleConnTimeout of 90s. Some of these tests rely on this timeout being
hit, meaning that there's a minimum time of 90s for these tests. This
resulted in the tsa suite being the slowest to complete, typically
taking ~5m.

This commit compiles the tsa binary with an overridden IdleConnTimeout
of 5s in tests. This does not affect the behaviour of Concourse, as the
default behaviour is left unchanged.

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>