atc: add interval to image check plans
We forgot to add the interval to the image check plans and this caused a
bug where image checks always ran. The interval on the check plan is
used to determine if we need to run if now - the last check time >
interval. If it has been less than the interval then we should not run,
this is so that we don't run a check for a custom type more than once
per interval.

Previously, we did not need to add the interval to the image check plans
because we only constructed an image check if the resource type did not
have a version (which should be only when it first runs in a pipeline).
Now that we always construct the image check, we need the interval to
limit the amount of times the check actually runs.

One thing to note is that image checks from fly execute builds will
always run. Previously, we did not always run a check if the image
alrady had a version but now that we don't have an interval set on the
plans constructed from fly execute, it will always run a check.

[#6274]

Co-authored-by: Esteban Foronda <eforonda@vmware.com>
Signed-off-by: Clara Fu <fclara@vmware.com>