atc: do not acquire checking lock if we do not need to run
Previously, we were first acquiring the checking lock and then seeing if
we need to run the check. This means that even if we know that we do not
need to run, it still holds the lock until the step completes. This
became a problem when we are creating checks for images and since a
custom type can be used by many resources across many jobs/pipelines,
the locking was causing some checks to take super long (15 minutes) to
acquire the lock.

So rather than acquiring the lock and then seeing if we need to run,
this commit changes it so that it will check to see if it needs to run
and only if it needs to run will it acquire the lock. It will do this
periodically every loop (1 second). I have slight concerns over the
amount of queries that will be run against the db because everytime it
loops for every resource/resource type/prototype, it will run a query to
fetch the last check times.

[#6274]

Signed-off-by: Clara Fu <fclara@vmware.com>