B: populate Elm feature flags on init
This fixes a bug with Causality where refreshing the page (or opening a
link to a Causality page) displays the "404 Not Found" page.

This was happening because we only learned about what feature flags were
active after the ClusterInfo response was returned, but we switched our
SubPage to NotFound before then (and the NotFound page doesn't handle
updates to ClusterInfo).

Instead of just fixing that, I took a different approach - populate the
feature flags on initialization using the Elm "flags" mechanism (which
is intended for things like this). This way, we don't need to wait for
the response, which simplifies things in Elm.

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>