remove unnecessary test
I think I was trying to test out the looping mechanic and how it would
recheck the last check value, but I did not write the test properly.
Instead, I think it's enough to just test that it checks the last check
value before we try to acquire the lock.

[#6274]

Signed-off-by: Clara Fu <fclara@vmware.com>