baggageclaim: prefer overlay to btrfs
when using the default driver "detect", previously we defaulted to using
btrfs if it was available on the machine.

https://github.com/concourse/concourse/discussions/5463 discusses the
trade-offs between drivers. in short:

* btrfs is faster when starting up privileged containers, but tends to
  be less stable, and may require loopback devices to work (which may be
  leaked)
* overlay is slow when starting up privileged containers but is more
  stable

however, since concourse/baggageclaim#44, the slow privileged container
startup with overlay is largely mitigated, so there's no real reason to
prefer btrfs over overlay (at least as a default)

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>