integration: wait for volumes to get GC'd
after upgrade/downgrade. this way, we won't try to use stale volumes
that are still present in the DB but not on the worker.

there are probably some enhancements we can make to Concourse to prevent
the need to do this (e.g. re-import base resource type volumes,
ephemeral check containers, etc.), but for now, let's just get these
tests passing again

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>
Co-authored-by: Rui Yang <ruiya@vmware.com>