atc: Prevent open redirect to other hosts
We fell victim to this behaviour: https://github.com/golang/go/issues/38642

The solution in this commit ensures that the URI we redirect to is
always a path. The TrimLeft() removes all prefixed forward slashes,
which covers the case where a redirect uri starts with two or more
slashse.

We then always add a single forward slash which will ensure that
ParseRequestURI() interprets the URI as a path.

Now if someone visits `http://ci.concourse-ci.org/sky/login?redirect_uri=//google.com/`
they'll end up redirected to `http://ci.concourse-ci.org/google.com` which
will return a 404 and redirect one more time to the homepage.

In case the behaviour of ParseRequestURI() changes or is smarter in the
future, added check to ensure that if either Host or Scheme is
populated that the URI is not used for redirecting. The tests have also
been updated so this change in behaviour will cause the tests to fail.

Signed-off-by: Taylor Silva <tsilva@pivotal.io>
Co-authored-by: Esteban Foronda <eforonda@vmware.com>