S: don't query full resource in check delegate
we found the resource query to be the most expensive in our stress test
environment (testing against the in-memory check builds - on master,
there are more expensive queries, but it's high up in the list). each
query performs a seq scan on resource_config_scopes. the query is still
very fast (hence the optimizer choosing to use a seq scan), but it's
very much in the hot-path so gets called frequently.

however, there's no need to query the whole resource - all we need it
for is to create a scope and to update it on the resource. for this, all
we need is the resource id and the (pipeline_id, name). if we reduce our
queries to this minimal set, we should make things more performant.

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>