atc: clear resource's config ids if config updates
when a resources config updates it will eventually point to a new
version history, which is tied to the resource through the config_id and
config_scope_id columns.

When a resource config is initally updated the relationship between the
new and old config is not made until a check for the resource is run.
This results in a period where the algorithm may run and use a version
from the previous resource's config.

This sql trigger will set the config id's to NULL for any resources who
has their config updated. If a resource's config is not updated then the
config id columns are left as-is.

Signed-off-by: Taylor Silva <tsilva@pivotal.io>