db: only clear config when source field changes
fixes a test related to the resource renaming feature.

Also updated the test description because the previous one didn't make
any sense to me.

When it said "resource exists but is disabled" I thought it meant the
active column was set to false in the resources table for that resource.
Based on my reading of the test this is not at all what it meant.

Signed-off-by: Taylor Silva <tsilva@pivotal.io>