on_error should not run the hook when err is retriable
on_error would run even when the error returned by wrapped by an
exec.Retriable{}, which happens when the
`CONCOURSE_ENABLE_RERUN_WHEN_WORKER_DISAPPEARS` is enabled. When that
feature is enabled all "core" steps (get, put, task, check,
set_pipeline) get wrapped by a RetryError() step.

The RetryError step actually returns the err from the child step but
simply wraps it in an exec.Retriable{}. In the engine.Run() there's a
check to see if any err is wrapped in a Retriable{}. If it is then the
step that returned the error doesn't receive a delegate.Finish() event
and the engine re-runs the step. That's how RetryError "loops". The
problem with this is that the on_error step was not aware of Retriable
errors and would run when it should not.

Now on_error knows about Retriable errors and won't run if it receives
one.

Signed-off-by: Taylor Silva <tsilva@pivotal.io>
Co-authored-by: Muntasir Chowdhury <mchowdhury@pivotal.io>