integration: bump timeout for volume gc
The integration tests in our ci deployment flake on waiting for the
volumes to gc in the upgrade/downgrade tests. When we run the tests
locally, it never flakes. Therefore we thought that possibly the reason
why it is flaky on ci is the extra load on the workers is causing the
volumes gc to take longer. We wanted to try allowing more time for the
volume gc to run for those tests and see if that fixes the flaky tests
in ci.

Signed-off-by: Clara Fu <fclara@vmware.com>