do deps bump on preset schedule
this is reverted from ignoring patch updates from ALL go libraries.

It'd be ideal if we can do an on-schedule group updates for concourse
deps i.e. updates once per week with all bumps in one PR. This is
currently impossible for dependabot.

There is alternative like [renovatebot], which allows group updates
and schedule presets.

Signed-off-by: Rui Yang <ruiya@vmware.com>