db: Refactor pipeline lifecycle query
I introduced a bug with an earlier refactor. Once I figured out the bug
I went back to the query and further refactored it. I think I ended up
with a clearer version of the query.

In the initial query the pipelines table was brought in via a second
LEFT JOIN, called p2. This confused me on my reading of the code today.
I was like "why is pipelines brought in again? This looks dumb!" and
there was no test guarding against this change (there is now with this
commit). I added back the second LEFT JOIN but called the second
pipelines table "parent" so its more obvious WHY the table is being
brought in again.

Overall the query is a lot safer. It now only picks up pipelines that
have a parent_job_id (it didn't before for all cases, pretty crazy!) and
further filters for other criteria from that starting point. I think
it's much easier to read now too.

Signed-off-by: Taylor Silva <tasilva@vmware.com>