atc/db: pause pipelines whose jobs have no builds
This commit contains a failing test.

I realized that a pipeline that was set but never ran, like a newly set
pipeline, would result in the pipeline being paused. I wrote a test and
made a change to the query to fix that.

I then realized that this change makes it so any pipeline that has a job
with zero builds impossible for the pauser to ever pause the pipeline. I
added a test for this and it is the failing test in this commit. I
currently can't think of a query to get around this issue. Depending on
discussion this test or the first test may be removed.

For the first test I added I see this as kinda an edge case. I feel like
it's extremely unlikely that someone will set a pipeline AND unpause it
at the exact time the pauser runs, which is currently once every 24hrs.
I ran into this when trying to setup an integration test for the pauser
and found that the pauser kept pausing my newly set pipeline before I
could trigger my pipelines job. That's the context on how I found this
issue.

Signed-off-by: Taylor Silva <tasilva@vmware.com>