integration test for pipeline pauser
WHAT A PAIN IT WAS TO WRITE THIS.

So first, getting faketime with postgres was an ADVENTURE. Faketime only
works the first time you bring prostgres up, after that it borks and
time goes back to now instead of whatever future/present time you set.
Therefore the test HAS to start in the past.

Second PIA was that having the db in the past breaks the component
runner because it uses the date from postgres and go's time.Now() to
deteremine if the interval has elapsed. So having postgres in the past
means EVERY component runs every 10s. I had to set the pauser interval
to a crazy high number to ensure it'll only run once on startup and then
not run again.

Also Go doesn't work with faketime so that's why postgres is using it
and not Concourse. Apparently Go does some assembly nonsense to figure
out the time instead of making kernel syscalls

tldr: time is really hard to fake

Signed-off-by: Taylor Silva <tasilva@vmware.com>