show var source error on resource and build page
On build page, add 'do' step to step trees so error event could
be linked to it correctly.

On resource page, the check was created async and hoping backend
to finish the check build. But if var source failed in first, the
check build was never be able to run the `check` script. Now after
the check request, ui will ask for check build plan explicitly to
get the errored build events.

Signed-off-by: Rui Yang <ruiya@vmware.com>