fix bug in automatic pipeline archiver
If the last build of the parent pipeline was a failed/errored build that
did not get to the set_pipeline step for the child pipelines, that
previously would have caused the pipeline archiver to archive the child
pipeline because the parent_build_id would be less than the
latest_completed_build_id for the job since failed/errored builds count
towards the latest_completed_build_id. So I changed the query to also
only count if the latest_completed_build_id is a successful build.

Also found another bug where the pipeline archiver was archiving
pipelines that have already been archived.

[#8167]

Signed-off-by: Clara Fu <fclara@vmware.com>