add fly clear-versions command
The clear-versions command can take in either a resource or a resource
type to clear the versions for. It will also print a warning message
with all the resources/resource types that the clearing of versions will
affect.

Also added the endpoints to the go-concourse package so that they can be
used by fly.

[#5418]

Signed-off-by: Clara Fu <fclara@vmware.com>