don't run resource type check if within check interval
We seemed to be running a check for resource types for every build
because we didn't have anything within the WaitToRun code for checking
if the check interval had elapsed.

This fix will help reduce the number of resource type checks, since it
is not really necessary to run a resource type check for every build.

[#8244]

Signed-off-by: Clara Fu <fclara@vmware.com>