add resource check before smoke tests
if a worker registered to ATC but not functional e.g. failed to run
baggageclaim activity the smoke tests will hang. This additional
resource check makes sure the worker is healthy. If not, it will fail
fast and also expose the worker error instead of just showing timing
out in the test.

Signed-off-by: Rui Yang <ruiya@vmware.com>