bump lager to v3
when bumping ginkgo to v2, a third party lib of ginkgo has to be
included in go.mod to fix a ginkgo flag redefined error. The root
cause is a test file imports ginkgo v2 and a lib who imports ginkgo
v1.

So now we need to bump lager to v3 who also uses ginkgo v2, together
with concourse/retryhttp and concourse/flag bumps to get rid of the
ginkgo v1 reference everywhere.

Signed-off-by: Rui Yang <ruiya@vmware.com>