correctly configure the garden backend in integration tests
We needed to call runtime.WithRequestTimeout() in order to correctly
configure the createContainer lock with a non-zero timeout. It's
actually amazing that all the other tests in here also weren't failing
in a similar manner.

I figured this out by putting `fmt.Println` statements everywhere and
printing out the configured timeout. In the lock, I was seeing that it
was configured with a `0s` timeout. I eventually realized that we were
setting the timeout on the containerd client, but not on the Garden
client (which contains the containerd client), and we create the lock
with the timeout from the Garden client, not the containerd client.
oops!

We configure the garden client correctly in the actual code, which is
why no user has ever had an issue with creating containers like we were
seeing in this test suite.

Signed-off-by: Taylor Silva <dev@taydev.net>