add "cgroup: host" to integration compose file
on hosts running cgroup v2 only, which is all of the OSS workers, we
need to start in the parent cgroup so we have access to all the
controllers. Otherwise docker starts us in a cgroup with no controllers
and we can only use controllers that are our parent cgroup passed down
to us.

Signed-off-by: Taylor Silva <dev@taydev.net>