properly close IO
After reviewing how others user the containerd client, I saw that no one
ever calls _just_ the `Wait()` func. They always call `Cancel()`,
`Wait()`, and `Close()` in that sequence. I think this was us improperly
using the containerd client.

I'm actually not even sure we need to call any of the functions on
`IO()` at all. After digging into the containerd client code I noticed
that the `process.Delete()` we call right after also calls those three
functions on the `IO` struct. So what we're doing might be completely
redundant!

I did test the issues in #8462 and #6775. I made sure I could get the
code in a state where I could reproduce both issues.

When I reverted the change made to fix #6775 I was still able to
reproduce the error reported in that issue. The bug in #8462 also
disappeared.

When I removed all of our calls to the `IO()` functions I was unable to
reproduce either of the bugs.

When I added the `Cancel()` and `Close()` functions I was again unable
to reproduce either of the bugs.

So I _think_ it would be fine to remove all of the `IO()` calls here. I
am erring on the side of caution though. I wasn't involved in the
initial implementation of the containerd runtime in Concourse so I don't
know why we make a call to manually close `IO()` at all. Reviewing git
history doesn't shed anymore light as to why.

Anyways, it seems fine for us to close IO on our own, as long as we do
it properly :)

Signed-off-by: Taylor Silva <dev@taydev.net>