rename interface Conn to DbConn
This interface is embedded in the LogConn struct. Because we expect the
interface to have a method named "Conn" this resulted in a naming
conflict when the interface is embedded because the method name now
conflicts with the embedded name of the interface... yeah that's
confusing to explain. Anyways, had to rename Conn so LogConn could
satisfy the Conn interface, that's the gist of this commit.

Signed-off-by: Taylor Silva <dev@taydev.net>