use github.com/google/uuid to generate UUIDv4
We found UUID collisions occuring in our cluster. We thought it was
impossible but I decided to look into the library we were using for UUID
generation and found this issue: https://github.com/nu7hatch/gouuid/issues/28

Which would explain some of the errors I've seen where the only possible
explanation was that we got the same UUID twice.

Signed-off-by: Taylor Silva <dev@taydev.net>