make container deletion more robust
We found that if a container wasn't correctly made, specifically if the
`/etc/hosts` file of the container wasn't populated with the containers
IP, container cleanup would fail with this error:

```
resume container traffic: error getting container IP: ip not found for container handle: 0ce036f8-aa2a-4bea-acbd-8ee48e5fa63b
```

We can ignore this error when deleting a container because it tells us
there are no iptables rules to remove anyways. We can probably continue
deleting the container and will delete it successfully.

Signed-off-by: Taylor Silva <dev@taydev.net>