don't create duplicate in-memory checks
Previously we could spawn as many in-memory checks and they wouldn't be
checked for duplication until the atc/db/lock acquired a Mutex and
tried to get a lock from the database.

I'm _guessing_, based on the info from
https://github.com/concourse/concourse/issues/8638 that the contention
for this lock would get really high. I'm guessing this and other users
cluster are very high usage clusters. It's possible that lidar may send
off multiple checks for the same resource which would result in multiple
go routines running and fighting over the same lock. Lidar would send
off multiple in-memory checks if every time it runs it still finds that
a resource has exceeded it's check interval.

I'm not 100% sure this is the problem and I don't have an easy way to
test this. This is all based on my reading of the code as it is today
and the pprof memory graph reported from users that showed a lot of
in-memory checks waiting on the db/lock Mutex.

Signed-off-by: Taylor Silva <dev@taydev.net>