exclude fakes from interface{} to any transition
These fakes are generated using `go generate ./...` by the counterfeiter
library https://github.com/maxbrunsfeld/counterfeiter

In the spirit of not fighting the tools we use, I want to exclude these
files from this change because the result is an endless fight of
flipping between `interface{}` and `any`.

When upstream counterfeiter starts generating fakes with `any`, then
great, we'll migrate the fakes then.

Signed-off-by: Taylor Silva <dev@taydev.net>