update README
Comment out the v10 block since it's stale and doesn't reflect the true
state of the project. Leaving it there for my own easy link reference
when planning the actual future of the project.

Removed the link to the Stark & Wayne tutorial since it's also stale now
and replaced it with the Getting Started guide on our website.

Signed-off-by: Taylor Silva <dev@taydev.net>