fix bug where Attach would fail if the worker restarted
Intergration test failed where we restart the worker and expect we
should be able to keep streaming a builds logs, but it failed with a
panic. Our Attach() expected that Creator would always be called before
Attach(), which would guarantee that the container's tasks were already
being tracked in the map[].

This is not the case if the worker is restarted though. The map[] is
empty in that case and we call Attach() since the container already
exists and is running. Added checks when reading the map[] so our
Attach() doesn't panic reading a nil map.

Signed-off-by: Taylor Silva <dev@taydev.net>