fix bug where re-attachment could fail when re-attachment previously occurred once already
bug was discovered when doing rolling updates on a cluster with two web
nodes. The first re-attchment would succeed, but the second would fail.
This appears to happen because we fully closed the previous CIO
(prevIO.Close()) which signled to containerd to delete the task's FIFO
files. It seems containerd eventually has a chance to delete the files
during the second re-attachment, which results in this error:

```
load proc: failed to open stdin fifo: error creating fifo /run/containerd/fifo/1731945221/task-stdin: no such file or directory
```

We avoid this by never calling `Close()` on any of the previous CIO's.
Calling `Cancel()` is enough to stop the previous CIO's from reading the
FIFO files and allowing the new CIO to read all the logs from the
containerd task.

Signed-off-by: Taylor Silva <dev@taydev.net>