add elm port to fetch hostname
this is so we could show the full URL to download fly in the
instructions. Thank you goes to this SO answer that provided a working
example: https://stackoverflow.com/a/60609998

Another popular SO answer that came up was this one: https://stackoverflow.com/a/51664724
But it appears to not work as expected anymore. The ports method felt
more reliable.

Signed-off-by: Taylor Silva <dev@taydev.net>