make task's working dir harder to determine
users could determine the working dir of the task because it was based
on the sha1 of the task name. The intention has always been to encourage
users to not rely knowing the working dir of their tasks.

This also removes usage of sha1 within the Go codebase of Concourse.

Signed-off-by: Taylor Silva <dev@taydev.net>