bugfix: vault userpass auth cannot be used
when trying to exchange the user / pass combination for a token, the userpass
auth expects the PUT to be done against a path that has the desired username
as its last element.

this is already done for okta and ldap, this commit enables the appending logic
that exists for okta and ldap to also be executed for the userpass engine.