change put.inputs to default to detect inputs
This should not be a breaking change for any known resource type. The
only case where I can see this breaking for someone is where a resource
type has been designed to expect volumes with specific names to be
mounted into the put step. None of the core resource-types do this and I
haven't seen any resource types in the wild do this either.

Signed-off-by: Taylor Silva <dev@taydev.net>