use yaml/v2 for vars package
I tried using the latest yaml/v4 package but ran into an issue.

The current code assumes that yaml.Unmarshal() will always return one of
three types:

- map[any]any
- []any
- string

If it's not any of those types then no vars get interpolated, which is
what I ran into. The v4 version will return maps and arrays of specific
types if the provided yaml is of that type. For example:

```yaml
key: value
```

Will be returned as map[string]any.

```yaml
key: value
1: 200
```

Will be returned as map[any]any beacuse the keys are mixed types. I
haven't tested with arrays but I assume it's a similar situation there.

This poses problems for our interpolation when the var is a different
type, like null. Tried using the reflect package to deal with this, but
in the end you have to deal with converting between different types and
for the null case I wasn't seeing a clear way to convert a string that's
a key into a null that is correctly parsed as yaml's null value.

Signed-off-by: Taylor Silva <dev@taydev.net>