combine upgrade and downgrade integration tests
The previous downgrade test was structured in a weird way. It was
testing a scenarion that I don't think would ever happen, which is
deploying the latest version (dev in this case) and then downgrading to
an older version. In reality, this would be like initally deploying 7.14
and then downgrading to 7.13. This is not a situation we support.

Now the test does the more common workflow of performing an upgrade and
then rolling back that upgrade. For example:

1. Deploy 7.13
2. Upgrade to 7.14
3. Downgrade to 7.13

Signed-off-by: Taylor Silva <dev@taydev.net>